refactor(message): move peer resolving logic to the peer package
